hw2/code_and_script/requirements.txt
transformers<=4.35.0.dev0
datasets==2.14.5
accelerate==0.23.0
sentencepiece==0.1.99
evaluate==0.4.0
rouge==1.0.1
spacy==3.7.1
nltk==3.8.1
ckiptagger==0.2.1
tqdm==4.66.1
pandas==2.1.1
jsonlines==4.0.0
protobuf==4.25.0